pyproject.toml
[project]
name = "llm-cache-router"
version = "0.2.1"
description = "Semantic cache + LLM provider router + cost tracker"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.27",
    "sentence-transformers>=3.0",
    "faiss-cpu>=1.8",
    "tenacity>=9.0",
    "numpy>=1.26",
]

[project.optional-dependencies]
redis = ["redis>=5.0"]
qdrant = ["qdrant-client>=1.9"]
fastapi = ["fastapi>=0.111", "starlette>=0.37"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]
all = ["llm-cache-router[redis,qdrant,fastapi]"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

